+++ src/helpers/module.ts
@@ -1,4 +1,5 @@
-import { StonexModule } from '../StonexModule'
+import { StoreBinder } from 'src/StoreBinder'
+import { PureStonexModule, StonexModule } from '../StonexModule'
 
 export function getAllMethodsFromModule (module: object): string[] {
   const methods: string[] = []
@@ -19,13 +20,12 @@ export function getAllMethodsFromModule (module: object): string[] {
     'toLocaleString',
   ]
 
-  const checkMethodOnExistInList = (value: any, key: string) =>
-    typeof value === 'function' &&
-    reservedMethods.indexOf(key) === -1 &&
-    methods.indexOf(key) === -1
-
   const addMethodToList = (key: string) => {
-    if (checkMethodOnExistInList(module[key], key)) {
+    if (
+      typeof module[key] === 'function' &&
+      reservedMethods.indexOf(key) === -1 &&
+      methods.indexOf(key) === -1
+    ) {
       methods.push(key)
     }
   }
@@ -45,11 +45,16 @@ export function isPureModule (module: any): boolean {
   return typeof module !== 'function'
 }
 
-export function convertToStandardModule (pureModule: object): any {
+/**
+ *
+ * @param {PureStonexModule<State>} pureModule
+ */
+export function convertToStandardModule<State = any, MP = any> (
+  pureModule: PureStonexModule<State>
+): any {
+  return class extends StonexModule<State> {
 
-  return class extends StonexModule<any> {
-
-    constructor (storeBinder: any) {
+    constructor (storeBinder: StoreBinder<State, MP>) {
       super(storeBinder)
 
       const keys = Object.keys(pureModule)